% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AP}
\alias{AP}
\title{Associated Press data}
\format{The data set is an object of class "\code{simple_triplet_matrix}" provided by package \strong{slam}. 
It is a word-document matrix which contains the term frequency of 7000 words in 2134 documents.}
\source{
Harman, D. (1992, November). Overview of the First Text REtrieval Conference (TREC-1). In TREC (Vol. 1992, pp. 1-20).
}
\usage{
data("AP")
}
\description{
Associated Press data from the First Text Retrieval Conference (TREC-1) 1992, 
which has being processed by stop-words removal, 
low-frequency words removal and short documents removal.
}
\keyword{datasets}
